ai_service/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-dotenv==1.0.1
langchain==0.3.26
langchain-google-genai==2.1.5
google-ai-generativelanguage==0.6.18
PyMuPDF==1.24.3 